requirements.txt
Quart==0.15 
pymongo==3.12.0
beautifulsoup4==4.9.3
playwright==1.40  # Add Playwright
python-dotenv==0.19.1
requests==2.25.1
openai==0.27.0
fuzzywuzzy==0.18.0
quart-cors==0.7.0 
python-Levenshtein==0.12.0
h11==0.14.0
hypercorn==0.15.0
wsproto==1.2.0
httpcore==1.0.2
httpx==0.25.2
jinja2==2.11.3
markupsafe==1.1.1
aiohttp==3.9.1
